Remove the re-export thing
apparently this is a syntax they're considering?
see gleam-lang - gleam - issues - 614
but I have genuinely no idea where I might've seen something that would
suggest to me that this was already a feature rofl

a previous version of this commit linked directly to the Gleam issue
there, but that leaves a note about my commit on the issue! I don't want
that lol

a second previous version linked to it less directly and that just added
ANOTHER mention to the issue lmfao. mortified. gleam people if you're
reading this im sorry